Remove StorageTask
